cross entropy

+++ pancreasCTDataset.py
@@ -70,7 +70,7 @@ class SplitTCIA3DDataset(data.Dataset):
         if self.hot == 1:
             target = self._toEvaluationOneHot(target)
         input = torch.from_numpy(input[None,:,:,:]).float()
-        target = torch.from_numpy(target).float()
+        target = torch.from_numpy(target).long()
